requirements.txt
openai==1.51.2
ollama==0.3.3
python-dotenv==1.0.1
pinecone==5.3.1
langchain-ollama==0.2.0
Flask==1.1.2
Flask-Cors==5.0.0